+++ deepchem/models/losses.py
@@ -1,3 +1,8 @@
+import math
+
+import torch
+
+
 class Loss:
     """A loss function for use in training models."""
 
@@ -463,6 +468,214 @@ class ShannonEntropy(Loss):
             return torch.mean(-inputs * torch.log(1e-20 + inputs), -1)
 
         return loss
+
+
+class MolGraphConvMILoss():
+    """
+    Losses used to maximize the mutual information between molecular graph encodings. Encodings can be generated from molecular graphs using a graph convolutional network, and the mutual information is calculated between subgraphs (local encodings) and the entire molecule (global encodings) in order to enable unsupervised training.
+    
+    This class supports only PyTorch.
+    
+    Parameters:
+    device: torch.device
+        Device to run the model on.
+    measure: str
+        The divergence measure to use for the unsupervised loss. Options are 'GAN', 'JSD',
+        'KL', 'RKL', 'X2', 'DV', 'H2', or 'W1'.
+    average_loss: bool
+        Whether to average the loss over the batch
+    """
+
+    def __init__(self, device, measure='JSD', average_loss=True):
+        self.device = device
+        self.measure = measure
+        self.average_loss = average_loss
+
+    def local_global_loss(self, l_enc, g_enc, index):
+        """
+        Compares the local and global encodings of molecular graphs and returns the loss between them based on the measure specified. Global in this case refers to the global encoding of the entire molecule, and local refers to the encoding of subgraphs.
+        
+        Parameters:
+        ----------
+        l_enc: torch.Tensor
+            Local feature map from the encoder.
+        g_enc: torch.Tensor
+            Global features from the encoder.
+        index: torch.Tensor
+            Index of the graph that each node belongs to.
+        device: torch.device
+            Device to run the model on.
+        measure: str
+            The divergence measure to use for the unsupervised loss. Options are 'GAN', 'JSD',
+            'KL', 'RKL', 'X2', 'DV', 'H2', or 'W1'.
+        average_loss: bool
+            Whether to average the loss over the batch
+
+        Returns:
+        -------
+        loss: torch.Tensor
+            Local Global Loss value
+        """
+        num_graphs = g_enc.shape[0]
+        num_nodes = l_enc.shape[0]
+
+        pos_mask = torch.zeros((num_nodes, num_graphs)).to(self.device)
+        neg_mask = torch.ones((num_nodes, num_graphs)).to(self.device)
+        for nodeidx, graphidx in enumerate(index):
+            pos_mask[nodeidx][graphidx] = 1.
+            neg_mask[nodeidx][graphidx] = 0.
+
+        res = torch.mm(l_enc, g_enc.t())
+
+        E_pos = self.get_positive_expectation(res * pos_mask, self.measure,
+                                              self.average_loss)
+        E_pos = (E_pos * pos_mask).sum() / pos_mask.sum()
+        E_neg = self.get_negative_expectation(res * neg_mask, self.measure,
+                                              self.average_loss)
+        E_neg = (E_neg * neg_mask).sum() / neg_mask.sum()
+
+        return E_neg - E_pos
+
+    def global_global_loss(self, g_enc, g_enc1):
+        """
+        Compares the encodings of two molecular graphs and returns the loss between them based on the measure specified. The encodings are generated by two separate encoders in order to maximize the mutual information between the two encodings.
+        
+        
+        Parameters:
+        ----------
+        g_enc: torch.Tensor
+            Global features from the encoder.
+        g_enc1: torch.Tensor
+            Global features from the separate encoder.
+        device: torch.device
+            Device to run the model on.
+        measure: str
+            The divergence measure to use for the unsupervised loss. Options are 'GAN', 'JSD',
+            'KL', 'RKL', 'X2', 'DV', 'H2', or 'W1'.
+        average_loss: bool
+            Whether to average the loss over the batch
+
+        Returns:
+        -------
+        loss: torch.Tensor
+            Global Global Loss value
+        """
+        num_graphs = g_enc.shape[0]
+
+        pos_mask = torch.eye(num_graphs).to(self.device)
+        neg_mask = 1 - pos_mask
+
+        res = torch.mm(g_enc, g_enc1.t())
+
+        E_pos = self.get_positive_expectation(res * pos_mask)
+        E_pos = (E_pos * pos_mask).sum() / pos_mask.sum()
+        E_neg = self.get_negative_expectation(res * neg_mask)
+        E_neg = (E_neg * neg_mask).sum() / neg_mask.sum()
+
+        return E_neg - E_pos
+
+    def get_positive_expectation(self, p_samples):
+        """Computes the positive part of a divergence / difference.
+    
+        Parameters:
+        ----------
+        p_samples: torch.Tensor
+            Positive samples.
+        average: bool
+            Average the result over samples.
+    
+        Returns:
+        -------
+        Ep: torch.Tensor
+            Positive part of the divergence / difference.
+        """
+        log_2 = math.log(2.)
+
+        if self.measure == 'GAN':
+            Ep = -torch.nn.functional.F.softplus(-p_samples)
+        elif self.measure == 'JSD':
+            Ep = log_2 - torch.nn.functional.F.softplus(-p_samples)
+        elif self.measure == 'X2':
+            Ep = p_samples**2
+        elif self.measure == 'KL':
+            Ep = p_samples + 1.
+        elif self.measure == 'RKL':
+            Ep = -torch.exp(-p_samples)
+        elif self.measure == 'DV':
+            Ep = p_samples
+        elif self.measure == 'H2':
+            Ep = 1. - torch.exp(-p_samples)
+        elif self.measure == 'W1':
+            Ep = p_samples
+        else:
+            raise ValueError('Unknown measure: {}'.format(self.measure))
+
+        if self.average_loss:
+            return Ep.mean()
+        else:
+            return Ep
+
+    def get_negative_expectation(self, q_samples):
+        """Computes the negative part of a divergence / difference.
+    
+        Parameters:
+        ----------
+        q_samples: torch.Tensor
+            Negative samples.
+        average: bool
+            Average the result over samples.
+    
+        Returns:
+        -------
+        Ep: torch.Tensor
+            Negative part of the divergence / difference.
+    
+        """
+        log_2 = math.log(2.)
+
+        if self.measure == 'GAN':
+            Eq = torch.nn.functional.F.softplus(-q_samples) + q_samples
+        elif self.measure == 'JSD':
+            Eq = torch.nn.functional.F.softplus(-q_samples) + q_samples - log_2
+        elif self.measure == 'X2':
+            Eq = -0.5 * ((torch.sqrt(q_samples**2) + 1.)**2)
+        elif self.measure == 'KL':
+            Eq = torch.exp(q_samples)
+        elif self.measure == 'RKL':
+            Eq = q_samples - 1.
+        elif self.measure == 'DV':
+            Eq = self.log_sum_exp(q_samples, 0) - math.log(q_samples.size(0))
+        elif self.measure == 'H2':
+            Eq = torch.exp(q_samples) - 1.
+        elif self.measure == 'W1':
+            Eq = q_samples
+        else:
+            raise ValueError('Unknown measure: {}'.format(self.measure))
+
+        if self.average_loss:
+            return Eq.mean()
+        else:
+            return Eq
+
+    def log_sum_exp(x, axis=None):
+        """Log sum exp function.
+    
+        Parameters
+        ----------
+        x: torch.Tensor
+            Input tensor
+        axis: int
+            Axis to perform sum over
+    
+        Returns
+        -------
+        y: torch.Tensor
+            Log sum exp of x
+    
+        """
+        x_max = torch.max(x, axis)[0]
+        y = torch.log((torch.exp(x - x_max)).sum(axis)) + x_max
+        return y
 
 
 def _make_tf_shapes_consistent(output, labels):
